Se organiza en codigo en distintos archivos para separar cada una de las clases y asi facilitar la lectura del codigo
