+++ main.py
@@ -1,0 +1,13 @@
+import sys
+import indicator_window
+from LedIndicatorWidget import *
+from PyQt5.QtWidgets import (QApplication)
+
+def main():
+    app = QApplication(sys.argv)
+    ex  = indicator_window.ExampleDisplay()
+    sys.exit(app.exec_())
+
+
+if __name__ == '__main__':
+    main()
